services/ingest-file/requirements.txt
banal>=0.4.2
normality>=1.0.0
pantomime>=0.3.3
servicelayer[google]==1.3.1
servicelayer[amazon]==1.3.1
balkhash[leveldb]==0.5.1
balkhash[sql]==0.5.1
followthemoney>=1.13.2
psycopg2-binary

grpcio>=1.21.1
google-cloud-vision==0.36.0

# Development
nose
click>=7.0

# File format support
dbf>=0.96.8
pdflib>=0.1.7
pymediainfo>=2.3.0
python-magic>=0.4.12
pypdf2>=1.26.0
rarfile>=3.0
xlrd>=1.2.0
pyicu>=2.0.3
openpyxl>=2.5.14
odfpy>=1.3.5
imapclient>=1.0.2
cchardet>=2.1.1
lxml>=4.2.1
olefile>=0.44
pillow>=5.1.0

# OCR tooling

# tesserocr>=2.2.2

# Get rid of:
# Used for LibreOffice converter service
cryptography==2.6.1
requests[security]>=2.21.0